data_generators/requirements.txt
kafka-python
Faker
scylla-driver